sound of translate, add suggest and accept suggest words

+++ app/src/main/java/com/example/projecttranslate/translate.java
@@ -3,12 +3,12 @@ package com.example.projecttranslate;
 public class translate {
     private String espanol;
     private String otomi;
-    private int id_traduccion;
+    private Long id_traduccion;
 
     public translate() {
     }
 
-    public translate(String espanol, String otomi, int id_traduccion) {
+    public translate(String espanol, String otomi, Long id_traduccion) {
         this.espanol = espanol;
         this.otomi = otomi;
         this.id_traduccion = id_traduccion;
@@ -30,11 +30,11 @@ public class translate {
         this.otomi = otomi;
     }
 
-    public int getId_traduccion() {
+    public Long getId_traduccion() {
         return id_traduccion;
     }
 
-    public void setId_traduccion(int id_traduccion) {
+    public void setId_traduccion(Long id_traduccion) {
         this.id_traduccion = id_traduccion;
     }
